It reads the length from the file, not array
