+++ Task2/HomeWork/Module.cpp
@@ -1,13 +1,6 @@
 #define _CRT_SECURE_NO_WARNINGS
 #include "Module.h"
 
-void swap(int &a, int&b)
-{
-	a = a ^ b;
-	b = a ^ b;
-	a = a ^ b;
-}
-
 int getLength()
 {
 	FILE *file = fopen("text.txt", "r");
@@ -48,62 +41,4 @@ int* initArray(int length)
 		newArray[i] = rand() % 100;
 	}
 	return newArray;
-}
-
-void quickSort(int* theArray, int firstIndex, int lastIndex)
-{
-	if (firstIndex >= lastIndex)
-	{
-		return;
-	}
-	int index = firstIndex;
-	int element = theArray[firstIndex];
-	for (int i = firstIndex; i <= lastIndex; ++i)
-	{
-		if (theArray[i] < element)
-		{
-			swap(theArray[i], theArray[index]);
-			++index;
-		}
-	}
-	if (index == firstIndex)
-	{
-		++index;
-	}
-	quickSort(theArray, firstIndex, index - 1);
-	quickSort(theArray, index, lastIndex);
-}
-
-int mostFrequentElement(int *theArray, int length)
-{
-	quickSort(theArray, 0, length - 1);
-	int element = 0;
-	int tmpElement = theArray[0];
-	int frequency = 0;
-	int tmpFrequency = 1;
-	for (int i = 1; i < length; ++i)
-	{
-		if (theArray[i] == tmpElement)
-		{
-			tmpFrequency++;
-		}
-		else
-		{
-			if (tmpFrequency > frequency)
-			{
-				element = tmpElement;
-				frequency = tmpFrequency;
-			}
-			tmpFrequency = 1;
-			tmpElement = theArray[i];
-		}
-	}
-	if (tmpFrequency > frequency)
-	{
-		return tmpElement;
-	}
-	else
-	{
-		return element;
-	}
 }
