+++ Task2/HomeWork/Module.h
@@ -1,11 +1,7 @@
 #pragma once
-#include <stdio.h>
-#include <time.h>
-#include <stdlib.h>
 
-void swap(int &a, int &b);
+#include "stdafx.h"
+
 int getLength();
 void printArray(int *theArray, int length);
 int* initArray(int length);
-void quickSort(int* theArray, int firstIndex, int lastIndex);
-int mostFrequentElement(int *theArray, int length);
