+++ Task2/HomeWork/SortAndFind.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "stdafx.h"
+
+void swap(int &a, int &b);
+void quickSort(int* theArray, int firstIndex, int lastIndex);
+int mostFrequentElement(int *theArray, int length);
