+++ Task2/HomeWork/Source.cpp
@@ -1,8 +1,8 @@
 #define _CRT_SECURE_NO_WARNINGS
-#include <stdio.h>
-#include <locale.h>
+#include "stdafx.h"
 #include "Module.h"
 #include "TestModule.h"
+#include "SortAndFind.h"
 
 int main()
 {
