+++ Task2/HomeWork/TestModule.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <stdio.h>
 #include "Module.h"
+#include "SortAndFind.h"
 
 void test();
