+++ Task2/HomeWork/stdafx.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <stdio.h>
+#include <time.h>
+#include <stdlib.h>
+#include <locale.h>
